Added casts due to the use of G++ flag -Wconversion

+++ code/boost_program_options_descriptions/src/CustomOptionDescription.cpp
@@ -68,7 +68,7 @@ namespace rad
   {
     for (size_t i = 0; i < positionalDesc.max_total_count(); ++i)
     {
-      if (optionID_ == positionalDesc.name_for_position(i))
+      if (optionID_ == positionalDesc.name_for_position((unsigned int)(i)))
       {
         boost::algorithm::erase_all(optionDisplayName_, "-");
         isPositional_ = true;
